some more javadoc

+++ data/crac/crac-api/src/main/java/com/farao_community/farao/data/crac_api/network_action/PstSetpoint.java
@@ -9,11 +9,20 @@ package com.farao_community.farao.data.crac_api.network_action;
 import com.farao_community.farao.data.crac_api.TapConvention;
 
 /***
+ * A PST setpoint is an Elementary Action which consists in changing
+ * the tap of a given PST.
+ *
  * @author Baptiste Seguinot {@literal <baptiste.seguinot at rte-france.com>}
  */
 public interface PstSetpoint extends ElementaryAction {
 
-    double getSetpoint();
+    /**
+     * Get the new setpoint (tap) that will be applied on the network element of the action
+     */
+    int getSetpoint();
 
+    /**
+     * Get the {@link TapConvention} with which the setpoint should be interpreted
+     */
     TapConvention getTapConvention();
 }
