+++ data/crac/crac-api/src/main/java/com/farao_community/farao/data/crac_api/network_action/PstSetpointAdder.java
@@ -17,7 +17,7 @@ public interface PstSetpointAdder {
 
     PstSetpointAdder withNetworkElement(String networkElementId);
 
-    PstSetpointAdder withSetpoint(double setPoint);
+    PstSetpointAdder withSetpoint(int setPoint);
 
     PstSetpointAdder withTapConvention(TapConvention rangeDefinition);
